:recycle: Fixes #211 -- ensure 24h time format field can be used
