+++ src/openforms/forms/static/forms/js/components/form/index.js
@@ -6,4 +6,3 @@ import './np-family-members';
 import './number';
 import './currency';
 import './email';
-import './time24h';
